+++ app/src/main/java/com/example/alessandro/apontemobile/modelo/Funcionario.java
@@ -4,9 +4,9 @@ import java.io.Serializable;
 import java.util.Date;
 
 public class Funcionario extends Pessoa implements Serializable {
-    private Long id;
+    private int id;
     private String cargo;
-    private float salario;
+    private double salario;
     private int hora_semanal;
     private int carteira_trabalho;
     private int pis;
@@ -19,9 +19,9 @@ public class Funcionario extends Pessoa implements Serializable {
     private Date entrada;
     private Date pagamento;
 
-    public Funcionario(String nome, String naturalidade, int celular, Date nascimento, String sexo, float altura,
-                       String logradouro, int numero, String bairro, String cep, String cidade, String uf, Long id,
-                       String cargo, float salario, int hora_semanal, int carteira_trabalho, int pis, int cpf,
+    public Funcionario(String nome, String naturalidade, int celular, Date nascimento, String sexo, double altura,
+                       String logradouro, int numero, String bairro, String cep, String cidade, String uf, int id,
+                       String cargo, double salario, int hora_semanal, int carteira_trabalho, int pis, int cpf,
                        int rg, int conta, String banco, int agencia, String camisa, Date entrada, Date pagamento) {
         super(nome, naturalidade, celular, nascimento, sexo, altura, logradouro, numero, bairro, cep, cidade, uf);
         this.id = id;
@@ -40,11 +40,11 @@ public class Funcionario extends Pessoa implements Serializable {
         this.pagamento = pagamento;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
@@ -56,11 +56,11 @@ public class Funcionario extends Pessoa implements Serializable {
         this.cargo = cargo;
     }
 
-    public float getSalario() {
+    public double getSalario() {
         return salario;
     }
 
-    public void setSalario(float salario) {
+    public void setSalario(double salario) {
         this.salario = salario;
     }
